数值警告剔除

+++ BabyT/Baby_TimeLineViewController.m
@@ -282,7 +282,7 @@
 -(void)onClickUser:(NSUInteger)userId
 {
     //点击左边头像 或者 点击评论和赞的用户昵称
-    NSLog(@"onClickUser: %ld", userId);
+    NSLog(@"onClickUser: %ld", (unsigned long)userId);
     
     UserTimelineViewController *controller = [[UserTimelineViewController alloc] init];
     [self.navigationController pushViewController:controller animated:YES];
